pyproject.toml
[tool.poetry]
name = "amlaith-generatediff"
version = "0.1.4"
description = ""
authors = ["Amlaith <burjooi@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.4.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff_script:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
